requirements.txt
torch
torchvision
albumentations
glob
scikit-learn
ruamel.yaml
PyYaml
opencv-python
tqdm
